Move validation into the base chooser modal implementation
